Consumer/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
python-dotenv==1.0.0
zipp==3.17.0

boto3~=1.34.11
botocore~=1.34.11